Add nextest in README and support in scripts/local_QA.sh
Add a note about `nextest` in the README.

The script now checks if nextest is installed. If it is installed
`cargo nextest run` is run rather than `cargo test`.

The former is nearly twice as fast. It is rather wasteful of processors
and memory. But it is an easy improvment.