Avoid unecessarily clobbering existing measurement table cache (#7)
Previously we wrote an empty file, overwriting the cache file.
This was only to test that it's writable. Then we computed a new
table, and overwrite the empty file.

Now we write a dummy file with a different name to test if we can
write to the directory. On success, we delete this file. This leave
an existing measurement table cache intact.

If the user made a mistake, and terminates the program, an existing
cache file will not have been clobbered.

Co-authored-by: John Lapeyre <jlapeyre@users.noreply.github.com>